Integrate Redis for background PDF processing and refactor job management
